duration date to component
