+++ apps/judicial-system/web/src/components/Table/index.ts
@@ -6,3 +6,5 @@ export { default as SortButton } from './SortButton/SortButton'
 export { default as TableContainer } from './TableContainer/TableContainer'
 export { default as TableHeaderText } from './TableHeaderText/TableHeaderText'
 export { default as TableSkeleton } from './TableSkeleton/TableSkeleton'
+export { default as DurationDate } from './DurationDate/DurationDate'
+export { getDurationDate } from './DurationDate/DurationDate'
